refactor: move "Domains" resources to separate project (#35)
