+++ src/VirusTotalCore.Domains/VirusTotalCore.Domains.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <Version>1.1.1</Version>
+    <Title>VirusTotalCore.Domains</Title>
+    <Authors>Kostiantyn Sharykin</Authors>
+    <Description>Domain lookup library for VirusTotal — standalone package for querying domain reports, comments and votes.</Description>
+    <Copyright>MIT License</Copyright>
+    <PackageLicenseUrl>https://github.com/hunterlan/VirusTotalCore/blob/main/LICENSE</PackageLicenseUrl>
+    <RepositoryUrl>https://github.com/hunterlan/VirusTotalCore</RepositoryUrl>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <LangVersion>12</LangVersion>
+    <TargetFramework>net8.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../VirusTotalCore.Common/VirusTotalCore.Common.csproj" />
+  </ItemGroup>
+
+</Project>
