+++ tests/VirusTotalCore.Tests/VirusTotalCore.Tests.csproj
@@ -33,6 +33,7 @@
       <ProjectReference Include="..\..\src\VirusTotalCore.IpAddresses\VirusTotalCore.IpAddresses.csproj" />
       <ProjectReference Include="..\..\src\VirusTotalCore.Urls\VirusTotalCore.Urls.csproj" />
       <ProjectReference Include="..\..\src\VirusTotalCore.Files\VirusTotalCore.Files.csproj" />
+      <ProjectReference Include="..\..\src\VirusTotalCore.Domains\VirusTotalCore.Domains.csproj" />
     </ItemGroup>
 
     <ItemGroup>
